[Feat] #260 LargerInfoView CollectionView Header 초기화
